Rename 4월 5일 - 그리디알고리즘/2437.cpp to 04월 05일 - 그리디알고리즘/2437.cpp
